Notify even if the iterable has been changed from the outside
If we remove an item from the iterable and remove the selection on the same element, we still want to get a change event for that particular item.
Generally speaking, we want to get notified no matter _how_ the item has been deselected (if it was only removed from selection or from the iterable itself).